Move from package .../evm.store.hedera to .../evm.store
